Split implementation of SFTP API out from its interface.
